Organize folder structure and sort experiences and educations
