+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import './App.css';
-import Main from './components/Main';
-import Skills from './components/Skills';
+import Main from './components/Main/Main';
+import Skills from './components/Skills/Skills';
 
 export default function App() {
