--- src/components/Section.tsx
@@ -1,5 +1,0 @@
-export default function Section() {
-    return (
-        <h1>Section</h1>
-    )
-}
